feat: personalAccount 송금하기 기능 완성

+++ server/src/main/java/com/b110/jjeonchongmu/domain/account/dto/TransferRequestDTO.java
@@ -16,5 +16,5 @@ public class TransferRequestDTO {
 	private AccountType toAccountType;
 	private Long toAccountId;
 	private String tradeDetail;
-	private Long transferAmount;
+	private Integer transferAmount;
 }
